+++ qingdaofu/TabBarController/产品/产品详情/发布人信息/AllEvaluationViewController.m
@@ -345,7 +345,7 @@
 - (void)refreshFooterOfAllEvaluation
 {
     _pageEva++;
-    NSString *page = [NSString stringWithFormat:@"%d",_pageEva];
+    NSString *page = [NSString stringWithFormat:@"%ld",(long)_pageEva];
     [self getEvaluateDetailListsWithPage:page];
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         [self.allEvaTableView footerEndRefreshing];
