+++ qingdaofu/TabBarController/消息/controller/系统消息/SystemMessagesViewController.m
@@ -203,7 +203,7 @@
 - (void)footerRefreshWithMessageOfSystem
 {
     _pageSys++;
-    NSString *page = [NSString stringWithFormat:@"%d",_pageSys];
+    NSString *page = [NSString stringWithFormat:@"%ld",(long)_pageSys];
     [self getSystemMessageListWithPage:page];
     
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
