+++ qingdaofu/TabBarController/用户/我的代理/controller/MyAgentListViewController.m
@@ -203,7 +203,7 @@
 - (void)footerRefreshOfMyAgent
 {
     _pageAgent ++;
-    NSString *page = [NSString stringWithFormat:@"%d",_pageAgent];
+    NSString *page = [NSString stringWithFormat:@"%ld",(long)_pageAgent];
     [self getMyagentListWithPage:page];
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         [self.myAgentListTableView footerEndRefreshing];
