+++ qingdaofu/TabBarController/用户/我的发布/处理进度/PaceViewController.m
@@ -216,7 +216,7 @@
 - (void)footerRefreshOfPace
 {
     _pagePace++;
-    NSString *page = [NSString stringWithFormat:@"%d",_pagePace];
+    NSString *page = [NSString stringWithFormat:@"%ld",(long)_pagePace];
     [self getPaceMessagesListWithPage:page];
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         [self.paceTableView footerEndRefreshing];
