+++ qingdaofu/TabBarController/用户/我的接单/处理中/view/TextFieldCell.m
@@ -117,7 +117,7 @@
     }
     
     _charCount= [[textView.text stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]]length];
-    self.countLabel.text = [NSString stringWithFormat:@"%d/600",_charCount];
+    self.countLabel.text = [NSString stringWithFormat:@"%ld/600",(long)_charCount];
     
     return YES;
 }
