+++ qingdaofu/TabBarController/用户/我的收藏/MyStoreViewController.m
@@ -200,7 +200,7 @@
 - (void)refreshFooterOfMySave
 {
     _pageStore ++;
-    NSString *page = [NSString stringWithFormat:@"%d",_pageStore];
+    NSString *page = [NSString stringWithFormat:@"%ld",(long)_pageStore];
     [self getMyStoreListWithPage:page];
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         [self.myStoreTableView footerEndRefreshing];
